Move general master data related rules
