+++ source/Energinet.DataHub.MeteringPoints.Domain/Energinet.DataHub.MeteringPoints.Domain.csproj
@@ -34,4 +34,8 @@ limitations under the License.
     <PackageReference Include="NodaTime" Version="3.0.5" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="MasterDataHandling\Components" />
+  </ItemGroup>
+
 </Project>
