+++ source/Energinet.DataHub.MeteringPoints.Domain/MasterDataHandling/Rules/CapacityFormatRule.cs
@@ -13,10 +13,9 @@
 // limitations under the License.
 
 using System;
-using System.Globalization;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Rules
+namespace Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Rules
 {
     public class CapacityFormatRule : IBusinessRule
     {
