+++ source/Energinet.DataHub.MeteringPoints.Domain/MasterDataHandling/Rules/InvalidMeterIdRuleError.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Rules
+namespace Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Rules
 {
     public class InvalidMeterIdRuleError : ValidationError
     {
