+++ source/Energinet.DataHub.MeteringPoints.Domain/MeteringPoints/Capacity.cs
@@ -14,6 +14,7 @@
 
 using System.Collections.Generic;
 using System.Globalization;
+using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Rules;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Rules;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
